Update main.c
It's the main source file for finding the histogram of an image. It contains the function call such as reading the image data, getting the histogram of the image read, writing an output image of the image read, and finally freeing the memory used to store the image read data.